As mentioned in my first mail, if the situation which you describe happens, the CopycallBack is not called at all. So no chance to interfere here.
It seems this is a bug in Windows: If desktop.ini contains a LocalizedResourceName then the only action during renaming by explorer is to adapt desktop.ini with the new name. But if the folder which contains desktop.ini is readonly, explorer simply does not call the registered CopyCallBack at all, but anyhow edits desktop.ini (this is weird)

ln.exe does not have the problem, because it does its job, without relying on callbacks from the shell-explorer beeing called.